fix: bank facade, matching pattern

+++ Banks/AccountCreation/CreditAccountCreator.cs
@@ -6,11 +6,11 @@ namespace Banks.AccountCreation
     public class CreditAccountCreator : IAccountCreator
     {
         private Client _client;
-        private int _limit;
+        private double _limit;
         private double _commission;
         private double _maxWithdrawalSum;
 
-        public CreditAccountCreator(Client client, int limit, double commission, double maxWithdrawalSum)
+        public CreditAccountCreator(Client client, double limit, double commission, double maxWithdrawalSum)
         {
             _limit = limit;
             _client = client;
